Extract cluster detail table into a component.
