+++ src/components/cluster_detail/index.js
@@ -15,12 +15,10 @@ import { organizationCredentialsLoad } from '../../actions/organizationActions';
 import * as clusterActions from '../../actions/clusterActions';
 import * as releaseActions from '../../actions/releaseActions';
 
-import { relativeDate } from '../../lib/helpers.js';
-import AWSAccountID from '../shared/aws_account_id';
 import Button from '../shared/button';
 import ClusterIDLabel from '../shared/cluster_id_label';
 import ClusterKeyPairs from './key_pairs';
-import ReleaseDetailsModal from '../modal/release_details_modal';
+import ClusterDetailTable from './cluster_detail_table';
 import ScaleClusterModal from './scale_cluster_modal';
 import UpgradeClusterModal from './upgrade_cluster_modal';
 
@@ -34,9 +32,6 @@ class ClusterDetail extends React.Component {
 
     if (props.cluster === undefined) {
       props.dispatch(push('/organizations/' + props.organizationId));
-      this.setState({
-        notfound: true,
-      });
 
       props.dispatch(
         flashAdd({
@@ -73,46 +68,9 @@ class ClusterDetail extends React.Component {
     }
   }
 
-  getMemoryTotal() {
-    if (!this.props.cluster.workers) {
-      return null;
-    }
-    var m = 0.0;
-    for (var i = 0; i < this.props.cluster.workers.length; i++) {
-      m += this.props.cluster.workers[i].memory.size_gb;
-    }
-    return m.toFixed(2);
-  }
-
-  getStorageTotal() {
-    if (!this.props.cluster.workers) {
-      return null;
-    }
-    var s = 0.0;
-    for (var i = 0; i < this.props.cluster.workers.length; i++) {
-      s += this.props.cluster.workers[i].storage.size_gb;
-    }
-    return s.toFixed(2);
-  }
-
-  getCpusTotal() {
-    if (!this.props.cluster.workers) {
-      return null;
-    }
-    var c = 0;
-    for (var i = 0; i < this.props.cluster.workers.length; i++) {
-      c += this.props.cluster.workers[i].cpu.cores;
-    }
-    return c;
-  }
-
   showDeleteClusterModal(cluster) {
     this.props.clusterActions.clusterDelete(cluster);
   }
-
-  showReleaseDetails = () => {
-    this.releaseDetailsModal.show();
-  };
 
   showScalingModal = () => {
     this.scaleClusterModal.getWrappedInstance().reset();
@@ -189,386 +147,144 @@ class ClusterDetail extends React.Component {
   };
 
   render() {
-    var instanceTypeOrVMSize = <tr />;
-    if (this.state.loading === false) {
-      if (this.props.provider === 'aws') {
-        instanceTypeOrVMSize = (
-          <tr>
-            <td>EC2 instance type</td>
-            <td className='value code'>
-              {this.props.cluster.workers[0].aws.instance_type}
-            </td>
-          </tr>
-        );
-      } else if (this.props.provider === 'azure') {
-        instanceTypeOrVMSize = (
-          <tr>
-            <td>VM size</td>
-            <td className='value code'>
-              {this.props.cluster.workers[0].azure.vm_size}
-            </td>
-          </tr>
-        );
-      }
-    }
-
-    var availabilityZonesOrNothing = null;
-    if (this.props.cluster.availability_zones) {
-      availabilityZonesOrNothing = (
-        <tr>
-          <td>Availablility zones</td>
-          <td className='value'>
-            {this.props.cluster.availability_zones.join(', ')}
-          </td>
-        </tr>
-      );
-    }
-
-    // BYOC provider credential info
-    var credentialInfoRows = [];
-    if (
-      this.props.cluster &&
-      this.props.cluster.credential_id &&
-      this.props.cluster.credential_id != '' &&
-      this.props.credentials.items.length === 1
-    ) {
-      // check if we have the right credential info
-      if (
-        this.props.credentials.items[0].id !== this.props.cluster.credential_id
-      ) {
-        credentialInfoRows.push(
-          <tr key='providerCredentialsInvalid'>
-            <td>Provider credentials</td>
-            <td className='value'>
-              Error: cluster credentials do not match organization credentials.
-              Please contact support for details.
-            </td>
-          </tr>
-        );
-      } else {
-        if (this.props.provider === 'aws') {
-          credentialInfoRows.push(
-            <tr key='awsAccountID'>
-              <td>AWS account</td>
-              <td className='value code'>
-                <AWSAccountID
-                  roleARN={
-                    this.props.credentials.items[0].aws.roles.awsoperator
-                  }
-                />
-              </td>
-            </tr>
-          );
-        } else if (this.props.provider === 'azure') {
-          credentialInfoRows.push(
-            <tr key='azureSubscriptionID'>
-              <td>Azure subscription</td>
-              <td className='value code'>
-                {
-                  this.props.credentials.items[0].azure.credential
-                    .subscription_id
-                }
-              </td>
-            </tr>
-          );
-          credentialInfoRows.push(
-            <tr key='azureTenantID'>
-              <td>Azure tenant</td>
-              <td className='value code'>
-                {this.props.credentials.items[0].azure.credential.tenant_id}
-              </td>
-            </tr>
-          );
-        }
-      }
-    }
-
     return (
-      <Breadcrumb
-        data={{
-          title: this.props.cluster.id,
-          pathname:
-            '/organizations/' +
-            this.props.cluster.owner +
-            '/clusters/' +
-            this.props.cluster.id,
-        }}
-      >
-        <Breadcrumb
-          data={{
-            title: this.props.cluster.owner.toUpperCase(),
-            pathname: '/organizations/' + this.props.cluster.owner,
-          }}
-        >
+      <div>
+        {this.state.loading === false ? (
           <Breadcrumb
-            data={{ title: 'ORGANIZATIONS', pathname: '/organizations/' }}
+            data={{
+              title: this.props.cluster.id,
+              pathname:
+                '/organizations/' +
+                this.props.cluster.owner +
+                '/clusters/' +
+                this.props.cluster.id,
+            }}
           >
-            <DocumentTitle
-              title={
-                'Cluster Details | ' + this.clusterName() + ' | Giant Swarm'
-              }
+            <Breadcrumb
+              data={{
+                title: this.props.cluster.owner.toUpperCase(),
+                pathname: '/organizations/' + this.props.cluster.owner,
+              }}
             >
-              {this.state.loading === false ? (
-                <div>
-                  <div className='cluster-details'>
-                    <div className='row'>
-                      <div className='col-7'>
-                        <h1>
-                          <ClusterIDLabel
-                            clusterID={this.props.cluster.id}
-                            copyEnabled
-                          />{' '}
-                          {this.props.cluster.name}{' '}
-                          {this.state.loading ? (
-                            <img
-                              className='loader'
-                              width='25px'
-                              height='25px'
-                              src='/images/loader_oval_light.svg'
-                            />
-                          ) : (
-                            ''
-                          )}
-                        </h1>
-                      </div>
-                      <div className='col-5'>
-                        <div className='pull-right btn-group'>
-                          <Button onClick={this.accessCluster}>
-                            GET STARTED
-                          </Button>
-                          {this.canClusterScale() ? (
-                            <Button onClick={this.showScalingModal}>
-                              SCALE
-                            </Button>
-                          ) : (
-                            undefined
-                          )}
-
-                          {this.canClusterUpgrade() ? (
-                            <Button onClick={this.showUpgradeModal}>
-                              UPGRADE
-                            </Button>
-                          ) : (
-                            undefined
-                          )}
-                        </div>
-                      </div>
-                    </div>
-                  </div>
+              <Breadcrumb
+                data={{ title: 'ORGANIZATIONS', pathname: '/organizations/' }}
+              >
+                <DocumentTitle
+                  title={
+                    'Cluster Details | ' + this.clusterName() + ' | Giant Swarm'
+                  }
+                >
                   <div>
                     <div className='cluster-details'>
                       <div className='row'>
-                        <div className='col-12'>
-                          <table className='table resource-details'>
-                            <tbody>
-                              <tr>
-                                <td>Created</td>
-                                <td className='value'>
-                                  {this.props.cluster.create_date
-                                    ? relativeDate(
-                                        this.props.cluster.create_date
-                                      )
-                                    : 'n/a'}
-                                </td>
-                              </tr>
-                              {credentialInfoRows === []
-                                ? undefined
-                                : credentialInfoRows}
-                              {this.props.release ? (
-                                <tr>
-                                  <td>Release version</td>
-                                  <td className='value code'>
-                                    <a onClick={this.showReleaseDetails}>
-                                      {this.props.cluster.release_version}{' '}
-                                      {(() => {
-                                        var kubernetes = _.find(
-                                          this.props.release.components,
-                                          component =>
-                                            component.name === 'kubernetes'
-                                        );
-                                        if (kubernetes) {
-                                          return (
-                                            <span>
-                                              &mdash; includes Kubernetes{' '}
-                                              {kubernetes.version}
-                                            </span>
-                                          );
-                                        }
-                                      })()}
-                                    </a>{' '}
-                                    {this.canClusterUpgrade() ? (
-                                      <a
-                                        onClick={this.showUpgradeModal}
-                                        className='upgrade-available'
-                                      >
-                                        <i className='fa fa-info-circle' />{' '}
-                                        Upgrade available
-                                      </a>
-                                    ) : (
-                                      undefined
-                                    )}
-                                  </td>
-                                </tr>
-                              ) : (
-                                <tr>
-                                  <td>Kubernetes version</td>
-                                  <td className='value code'>
-                                    {this.props.cluster.kubernetes_version !==
-                                      '' &&
-                                    this.props.cluster.kubernetes_version !==
-                                      undefined
-                                      ? this.props.cluster.kubernetes_version
-                                      : 'n/a'}
-                                  </td>
-                                </tr>
-                              )}
-                              <tr>
-                                <td>Kubernetes API endpoint</td>
-                                <td className='value code'>
-                                  {this.props.cluster.api_endpoint
-                                    ? this.props.cluster.api_endpoint
-                                    : 'n/a'}
-                                </td>
-                              </tr>
-                              {availabilityZonesOrNothing}
-                              <tr>
-                                <td>Number of worker nodes</td>
-                                <td className='value'>
-                                  {this.props.cluster.workers
-                                    ? this.props.cluster.workers.length
-                                    : 'n/a'}
-                                </td>
-                              </tr>
-                              {instanceTypeOrVMSize}
-                              <tr>
-                                <td>Total CPU cores in worker nodes</td>
-                                <td className='value'>
-                                  {this.getCpusTotal() === null
-                                    ? 'n/a'
-                                    : this.getCpusTotal()}
-                                </td>
-                              </tr>
-                              <tr>
-                                <td>Total RAM in worker nodes</td>
-                                <td className='value'>
-                                  {this.getMemoryTotal() === null
-                                    ? 'n/a'
-                                    : this.getMemoryTotal()}{' '}
-                                  GB
-                                </td>
-                              </tr>
-                              {this.props.provider === 'kvm' ? (
-                                <tr>
-                                  <td>Total storage in worker nodes</td>
-                                  <td className='value'>
-                                    {this.getStorageTotal() === null
-                                      ? 'n/a'
-                                      : this.getStorageTotal()}{' '}
-                                    GB
-                                  </td>
-                                </tr>
-                              ) : (
-                                undefined
-                              )}
-                              {this.props.cluster.kvm &&
-                              this.props.cluster.kvm.port_mappings ? (
-                                <tr>
-                                  <td>Ingress Ports</td>
-                                  <td>
-                                    <dl className='ingress-port-table'>
-                                      {this.props.cluster.kvm.port_mappings.reduce(
-                                        (acc, item, idx) => {
-                                          return acc.concat([
-                                            <dt key={`def-${idx}`}>
-                                              <code>{item.protocol}</code>
-                                            </dt>,
-                                            <dd key={`term-${idx}`}>
-                                              {item.port}
-                                            </dd>,
-                                          ]);
-                                        },
-                                        []
-                                      )}
-                                    </dl>
-                                  </td>
-                                </tr>
-                              ) : (
-                                undefined
-                              )}
-                            </tbody>
-                          </table>
+                        <div className='col-7'>
+                          <h1>
+                            <ClusterIDLabel
+                              clusterID={this.props.cluster.id}
+                              copyEnabled
+                            />{' '}
+                            {this.props.cluster.name}{' '}
+                            {this.state.loading ? (
+                              <img
+                                className='loader'
+                                width='25px'
+                                height='25px'
+                                src='/images/loader_oval_light.svg'
+                              />
+                            ) : (
+                              ''
+                            )}
+                          </h1>
+                        </div>
+                        <div className='col-5'>
+                          <div className='pull-right btn-group'>
+                            <Button onClick={this.accessCluster}>
+                              GET STARTED
+                            </Button>
+                            {this.canClusterScale() ? (
+                              <Button onClick={this.showScalingModal}>
+                                SCALE
+                              </Button>
+                            ) : (
+                              undefined
+                            )}
+
+                            {this.canClusterUpgrade() ? (
+                              <Button onClick={this.showUpgradeModal}>
+                                UPGRADE
+                              </Button>
+                            ) : (
+                              undefined
+                            )}
+                          </div>
                         </div>
                       </div>
                     </div>
+                    <div>
+                      <ClusterDetailTable
+                        canClusterUpgrade={this.canClusterUpgrade()}
+                        showUpgradeModal={this.showUpgradeModal}
+                        cluster={this.props.cluster}
+                        provider={this.props.provider}
+                        credentials={this.props.credentials}
+                        release={this.props.release}
+                      />
+                      <ClusterKeyPairs cluster={this.props.cluster} />
 
-                    <ClusterKeyPairs cluster={this.props.cluster} />
-
-                    <div className='row section cluster_delete col-12'>
-                      <div className='row'>
-                        <h3 className='table-label'>Delete This Cluster</h3>
+                      <div className='row section cluster_delete col-12'>
+                        <div className='row'>
+                          <h3 className='table-label'>Delete This Cluster</h3>
+                        </div>
+                        <div className='row'>
+                          <p>
+                            All workloads on this cluster will be terminated.
+                            Data stored on the worker nodes will be lost. There
+                            is no way to undo this action.
+                          </p>
+                          <Button
+                            bsStyle='danger'
+                            onClick={this.showDeleteClusterModal.bind(
+                              this,
+                              this.props.cluster
+                            )}
+                          >
+                            Delete Cluster
+                          </Button>
+                        </div>
                       </div>
-                      <div className='row'>
-                        <p>
-                          All workloads on this cluster will be terminated. Data
-                          stored on the worker nodes will be lost. There is no
-                          way to undo this action.
-                        </p>
-                        <Button
-                          bsStyle='danger'
-                          onClick={this.showDeleteClusterModal.bind(
-                            this,
-                            this.props.cluster
-                          )}
-                        >
-                          Delete Cluster
-                        </Button>
-                      </div>
-                    </div>
-                    <ScaleClusterModal
-                      ref={s => {
-                        this.scaleClusterModal = s;
-                      }}
-                      cluster={this.props.cluster}
-                    />
-
-                    {this.props.targetRelease ? (
-                      <UpgradeClusterModal
+                      <ScaleClusterModal
                         ref={s => {
-                          this.upgradeClusterModal = s;
+                          this.scaleClusterModal = s;
                         }}
                         cluster={this.props.cluster}
-                        release={this.props.release}
-                        targetRelease={this.props.targetRelease}
                       />
-                    ) : (
-                      undefined
-                    )}
+
+                      {this.props.targetRelease ? (
+                        <UpgradeClusterModal
+                          ref={s => {
+                            this.upgradeClusterModal = s;
+                          }}
+                          cluster={this.props.cluster}
+                          release={this.props.release}
+                          targetRelease={this.props.targetRelease}
+                        />
+                      ) : (
+                        undefined
+                      )}
+                    </div>
                   </div>
-                  {this.props.release ? (
-                    <ReleaseDetailsModal
-                      ref={r => {
-                        this.releaseDetailsModal = r;
-                      }}
-                      releases={[this.props.release]}
-                    />
-                  ) : (
-                    undefined
-                  )}
-                </div>
-              ) : (
-                <div className='app-loading'>
-                  <div className='app-loading-contents'>
-                    <img
-                      className='loader'
-                      src='/images/loader_oval_light.svg'
-                    />
-                  </div>
-                </div>
-              )}
-            </DocumentTitle>
+                </DocumentTitle>
+              </Breadcrumb>
+            </Breadcrumb>
           </Breadcrumb>
-        </Breadcrumb>
-      </Breadcrumb>
+        ) : (
+          <div className='app-loading'>
+            <div className='app-loading-contents'>
+              <img className='loader' src='/images/loader_oval_light.svg' />
+            </div>
+          </div>
+        )}
+      </div>
     );
   }
 }
